pyproject.toml
[project]
name = "ai-conversation-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.43",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "openai>=1.65.5",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
]
